lab02

+++ lab2/DebugExercise/DebugExercise1.java
@@ -1,12 +1,14 @@
 package DebugExercise;
 
+import java.awt.geom.Arc2D;
+
 /**
  * Exercise for learning how the debug, breakpoint, and step-into
  * feature work.
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient =(float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
